refactor(ui): move initialization functions into utils
